[LogAnalysis] Do not retry on certain error
After setting the percentage to 100%, we are seeing some tasks got stuck  https://screenshot.googleplex.com/8pK4ndStwViAp5r.

They all seem to belong to the same set of zephyr builders [1]. The error was that  the return data (~5MB) was bigger than capacity.

This is strange because we set the max size to download batch to 2MB [2].

My theory is that somehow for that builder, the artifact size was set wrongly, but I am not sure, and it's hard to debug which artifact causes the problem, because we do things in batch.

But it seems like when we encounter those error, we should not retry.

[1] https://docs.google.com/document/d/1YUBPapk5G6IDnzs4XNj_x98VOX2Anj6HLbnB5FJQPSQ/edit?tab=t.0

[2] https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/resultdb/internal/services/artifactexporter/artifact_exporter.go;l=68?q=artifactexporter


Bug:b/334684340
Change-Id: Ieb3a7605d18d6a04bec675f7635255537105fa10
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5508175
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>