[luci] Make Batch and other interfaces generic in sync/dispatcher
Address a comment raised in
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5512990
and make the Batch and SendFn and other interfaces generic as well, and
then deal with all of the pain and compile-time errors that this
causes.

This change ends up being pretty big and I don't see an obvious way to
keep it smaller, so I'll keep it distinct from its parent CL.

That way we can land the cool new Channel[T] first, and then add the
other things.

BUG=338472916
TEST=make test

Change-Id: I22ff297e98d241c693884c2617f710e4b0d99d04
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5514102
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>