[testing/assert] Add cmp.Diff integration, diff highlighting.
This also adds the ability to mark a Finding as 'verbose', which
will be summarized unless `-v` is passed to `go test` (the default
in CI).

This should obviate the need for the structureddiff library.

Integration with cmp.Diff is only done at the FailureBuilder level,
and highlighting for the diff itself is decoupled from this.

R=gregorynisbet@google.com

Bug: 319730573
Change-Id: If1a84a671b1658c13b4861c7b267dd43d3259cd7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5517134
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>