[swarming] Ignore STREAM_ALREADY_COMMITTED error.
It appears (based on logs on chrome-swarming) that GetWriteStream
is "eventually consistent", i.e. it may indicate the stream is
pending even after BatchCommitWriteStreams has successfully
finished (even if all calls happen serially, no race conditions).

When this happens and we try to commit the stream, we get back
STREAM_ALREADY_COMMITTED error, which marks the TQ task as
fatally failed (which is technically fine, but noisy).

Ignore this error. It indicates the retry wasn't needed and
everything is OK already.

R=chanli@chromium.org

Change-Id: I5e147d9b287dbfcccfa88340c70de114b832c584
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5520046
Reviewed-by: JB Tamas-Leloup <jtamasleloup@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>