[ResultDB] Rename QueryRunTestVariants to QueryRunTestVerdicts.
The new RPC name better aligns with the current state of terminology.

With this change, we also use a new RunTestVerdicts proto for
run test verdicts. A fit-for-purpose proto is cleaner and
gives more flexibility as the API surface evolves.

The RPC is not yet in use, so this substitution is still safe to
make.

BUG=b:333652213
TEST=Integration tests

Change-Id: I26492d24c11840c147c580fbb67e0fb09b27c562
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5523656
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>