[gce-provider] Remove existing bot functionality from manage-bot queue
This patch removes the manageExistingBot functionality from the queue.
This should reduce the time taken for this queue significantly. This
functionality is provided by the new deleteStaleSwarmingBots. And as it
is a separate queue this distributes the task enough to avoid the pileup
on the queue.

Change-Id: I1251fa9234ace479aa9719b49acc088fdc31ec3f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5512998
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>