[ftt] Add "from the top" replacement for Convey suites.
This implements a semantically-equivalent variant of the goconvey
"Convey" construct, except that:
  * This uses go testing's `t.Run` subtest mechanism
  * No usage of `gls` - no hidden state.
  * Fully compatible with all assertion libraries using `testing.TB`
    or `*testing.T`.
  * Supports parallel execution of sub-tests (goconvey does not).

This will allow us to programmatically rewrite our tests to drop the
dependency on goconvey (and thus gls) without having to manually touch
everything.

After dropping the goconvey dependency, we will be in a position
where tests may or may not be written with `ftt`, and it will be easy
to adapt them to/from this style vs plain tests just directly using
`t.Run()` if desired, because it will fully decouple the assertions
from the 'layout' of the tests.

R=gregorynisbet@google.com

Change-Id: Iab9cdc80a9fc90bccfe902a2bcfa2b7386f9ae8a
Bug: 319730573
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4917850
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>