[swarming] Refactor some RPC tests, in particular TestCountTasks.
This is in preparation for adding ListTasks.

The test was weird and didn't test a lot of cases. Make it
similar to other tests (in particular bot counting test) and
prepare for adding ListTasks RPC: its test will use the exact
same mocked tasks as CountTasks test (similar to how bot
counting and listing tests share the same state).

To make this "duality" more pronounced, move state mocking
functions into the shared testing utils file (base_test.go).
Share test time as well, since SetupTestTasks uses it.

Fix Internal error when time range is empty. It should be
InvalidArgument error.

Uncover a bug in handling of QUERY_PENDING_RUNNING filter. Its
use of "<=" datastore filter is incompatible with filtering by
time: a datastore query can have comparison filter on at most
one field.

The bug will be fixed in a separate CL. It is not trivial (we'll
need to replace "state <= PENDING" with "state in [PENDING, RUNNING]"
which means we may need to run multiple queries in parallel and
merge their results, just like we do when filtering on tags with
"|" construct in them).

R=chanli@chromium.org, jtamasleloup@google.com

Change-Id: I653d5e345b3c8925fbdba9ecfd48db8a5c944931
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5528117
Reviewed-by: JB Tamas-Leloup <jtamasleloup@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>