[swarming] Fix QUERY_PENDING_RUNNING handling in CountTasks.
It was adding an "<=" filter, which is not compatible when also
filtering by a timestamp (a datastore query can have at most one
property used in inequality filters).

We now run either one `state IN [PENDING, RUNNING]` query,
or two `state == PENDING || state == RUNNING` queries (if `IN` filter
is not allowed, which happens locally).

Furthermore, if we end up using the `IN` filter, all subsequent
query splitting (in particular when filtering on tags with "|" in
them), must not use `IN` datastore filters, since there can be at
most one `IN` filter per query. Force them to do full splitting.

R=chanli@chromium.org, jtamasleloup@google.com

Change-Id: I2d17fd74e9d1d18f58ad527346b2cbe98b557598
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5527582
Reviewed-by: JB Tamas-Leloup <jtamasleloup@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>