[luci-channel] Add ConsumeChannel to dispatcher.
Extract the channel consumer implementation from the CL below. Factor
it into a method Consume, which just reads from a standard channel and
ConsumeChannel, which constructs a dispatcher.Channel on the spot,
feeds a standard channel into it, and then hands control back to the
caller once it is drained.

BUG=None
TEST=make test

Change-Id: Ic3d3b91c67854295965b815e9676b9797cb7374b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5535673
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>