[auth-service] Remove perms match prereq for UpdateAuthRealmsGlobals
As part of switching the update-realms-cron to be done by Auth Service
v2, we need to remove this prereq of the permissions.cfg data matching
what is stored by Auth Service v1.

When we do the switch, v1 will no longer be maintaining the permissions.

Bug: b/321019030, b/40228798
Change-Id: Ia57dde239296749ea3e5f5b251d38ed96cbe5845
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5543741
Reviewed-by: Rachael Newitt <renewitt@google.com>
Commit-Queue: Anne Redulla <aredulla@google.com>