[testing/truth] Support adding source context for assertions inside helpers.
This solves the problem that `t.Helper()` obscures the line of an
actual assertion within a helper with multiple assertions, and
allows helper authors to annotate failing assertions with the line
number inside of the helper, in addition to the 'root' call to the
helper.

Example:

 func TestThing(t *testing.T) {
   myHelper := func(actual, expected myType) {
     // makes Go 'testing' package skip this to find original call.
     t.Helper()

     // We add WithLineContext to these comparisons so that if they
     // fail, we will see the file:line within this helper function.
     check.That(t, actual.field1,
         should.Equal(expected.field1).WithLineContext())

     check.That(t, actual.field2,
         should.Equal(expected.field2).WithLineContext())  // line YY
  }
  // ...
  myHelper(a, expected)  // line XX
 }

In this example, the test will output something like:

 --- FAIL: FakeTestName (0.00s)
     example_test.go:XX: Check should.Equal[int] FAILED
        (at example_test.go:YY)
        Actual: 10
        Expected: 20

Where XX is the line of the myHelper call, and YY is the line of the
actual should.Equal check inside of the helper function.

R=gregorynisbet@google.com

Bug: 319730573
Change-Id: Id836e47ac95d3a7b5c69efaba559e18a8b32bb2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5543543
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>