[Buildbucket] Add CustomMetrics in Build model
I expect most builders don't have custom metrics set up, so if we only store the custom metrics in the builder, we'll have to do a lot of unnecessary builder reads.

Instead we should save the possible custom metrics a build may report to at its creation time, and check it at build events to determine if we should report the build to some custom metrics.

This CL adds the field to Build proto, write and use it would be in follow-up CLs.

Bug: 338071541
Change-Id: I2dae0c89ad75ba95164e6ebf4c3a411c13120e16
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5544936
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>