[Analysis] Identify the changepoint being queried for culprit analysis.
Identify the range of commit positions we should retrieve and feed
into the analysis.

In a follow-up, we will query the test verdicts table to identify
the commit hash of the end commit in the range, and then we can go
to gitiles to retrieve the details of CLs involved.

BUG=b:343091350
TEST=Integration tests

Change-Id: Ic317d29a874505d21e1961347b35663632a346a1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5568638
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>