[swarming] Refactor acls.Checker to have a cache.
That way repeatedly calling CheckTaskPerm in a loop is
relatively fast and BatchCheckTaskPerm is not needed.

This was the original plan (as indicated in the Checker
comments), but I forgot about it.

Note that acls.Checker is scoped to a single request only,
so it is fine to memoize the outcomes "forever" (this
cache will be dropped very soon, once the request finishes).

Also note that acls.Checker currently is always used in the
main request goroutine, non-concurrently. We can avoid doing
any locking to further reduce performance difference between
repeated calls to CheckTaskPerm and a single (potentially
optimized) BatchCheckTaskPerm.

R=chanli@chromium.org

Change-Id: I85eed95ff95a18623222741ffff7a5d9d743e63d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5595322
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>