[Reproduction] Add button to step and test variant to indicate there is an instruction.
If a step or test variant have instruction, add an icon next to it, and when the user clicks on it, show a dialog (currently only empty - not implemented yet).

This will not show in prod because we have no instruction data there. The instruction was showing in the demo build because I added some data there.

Bug:b/345308119
Change-Id: Ia9ab064b175d36535dac0faf39b8574a51c9f838
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5602070
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>