[exception-db] Add new stream flag for uploading extended_properties
We can use the flag "inv-extended-properties-dir" to pass a directory
path containing files for the invocation's extended_properties in
JSON format.

Only files directly under this dir with the extension ".jsonpb" will be
read. The filename after removing ".jsonpb" and the file content will be
added as a key-value pair to the invocation's extended_properties map.

Bug: 337058844
Change-Id: I3f5ab66d42a05abebdaf6dc9977ad7ee188e88f4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5590629
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>