[prpc] Allow "peeking" into requests in the override callback.
The override callback is primarily used to implement traffic
splitting. It is called before an RPC handler and it may decide
to fully execute the request itself (usually by proxying it
to another service).

In Swarming we'll need to look at the pagination cursors to
decide if a listing request should be sent to Python or be
processed in Go. Thus the override callback needs a way to
"peek" into requests.

One complication is that http.Request.Body can be read at most
once. If we read it, and then decided to resume request
execution normally, we need to make sure it can be read again.
The only way to do that is to buffer the request in memory.
This should be fine, since most RPC requests are tiny (and
this is "temporary" code anyway).

For the same reason, the "peeking" is not optimal: if we resume
the processing of the request after "peeking", the request will
be deserialized again (while in theory we could cache the peeked
value and reuse it). But this will complicate the code even more,
and, as mentioned, requests are tiny and this is temporary.

R=chanli@chromium.org

Change-Id: I03d4691db78b39bd1e2fe5a37ef279d2e2226675
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5599264
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>