[swarming] Implement RPC traffic splitting.
Go server can now proxy a subset of requests to the Python
implementation. It decides based on request headers and based
on `traffic_migration` section in settings.cfg.

This is not enabled yet for real because there's no
dispatch.yaml rule that sends traffic to the Go server in the
first place (it will be added later once this implementation
is deployed, it will be safer to enable it separately).

The exact routing rules are:
1. `X-Route-To: go` header => handle in Go.
2. `X-Route-To: py` header => handle in Py.
3. `Host` header is "default-go-..." => handle in Go.
4. Route randomly based on configured `route_to_go_percent`.

Request proxied to python will have "/python/..." prepended to
their URL. That allows them to be dispatched to Python via
dispatch.yaml rules (instead of being sent back to Go).

When proxying requests to Py, appends `X-Routed-From-Go: 1`
header. If an incoming request has this header already
(which should never happen normally), doesn't proxy it
again. This is to prevent accidental endless proxying loop
in case of bugs (the way GAE works, it can be pretty
catastrophic).

R=chanli@chromium.org

Change-Id: I841861e87fd69d2865363cc13929b86658121352
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5606248
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>