[milo] Switch to v2 Swarming client.
This code has no tests, so it is possible this CL is buggy. But
IIUC this part of Milo is not very actively used and at worst
it will break some old LED (and maybe Skia) clients.

I'll try to verify on -dev after landing.

R=weiweilin@google.com, mwarton@google.com
BUG=b/295054045

Change-Id: I8cf8f31a8626ccfbc4668e01cf8ee45c42b2a1fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5606462
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Weiwei Lin <weiweilin@google.com>