luci-ui: improve CategoryTree API
Make the API closer to an ES6 Map. Comparing to an ES6 Map, CategoryTree
maps a key list to an item instead of a key to an item.

Notable changes:
1. Each node can only attach one value instead of an array of values.
   If more than one is required, the parent can append a sequential key
   to the category.
2. Rename .addValue to .set.
3. Add delete method.
4. Rename .entries to .enumerate since it does not return keys only
   indices.
5. Add .probe to support incrementally building the tree.

This will be used to build the ancestor->child build tree.

R=beining@google.com

Bug: b/345367426
Change-Id: I527e16157f91fcddd8b5d30d1872a3b958a4029b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5620050
Reviewed-by: Beining Chen <beining@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>