luci-ui: switch to use the new <BuildIdBar />
Merge this only when we want to release the redesigned <BuildIdBar />

Also changed <CustomBugLink /> to <BuildBugButton /> and houses the
button itself. This fixes the new <BuildIdBar /> displays an empty
button when there's no custom bug link. The change could not be made
easily in a separate CL because the old build bar still needs the
component to be a link.

R=mwarton@google.com

Bug: b/346868048
Change-Id: Ide5a2cef5bd0d76e6ac16a4d650e285e1c83fb87
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5632612
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>