[Analysis] Partition verdicts by invocation creation time.
For the moment, this seems to be the best approach. We lose
the guarantee that all of a CV run is in the same partition, but
we gain consistency across our datasets - from changepoint analysis,
to test results, to test verdicts and clustered failures.

BUG=None
TEST=Integration tests

Change-Id: I0eab1e544ee213ebbe6fa0fbbe1c9ebac39d5198
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5644587
Auto-Submit: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>