[swarming] Refactor `monitor` implementation.
This just moves code around and adds interface boundaries.
The only non-trivial change is parallel `Finalize(...)`
calls (will be important later).

The primary motivation is to make `monitor` more extendable.
In particular, there are 4 mechanisms in Swarming that need to
periodically visit every bot. Instead of 4 separate crons each
issuing datastore queries, we can have one query with 4
"visitors" passed to it. Existing `report-bots` cron will do
that.

Refactor active tasks scanner to be structured the same way
for consistency (it isn't hard). There's currently no plan to
extend it though.

Another goal of this refactoring is to move code out of
the giant main.go and make it more modular and testable.
All missing tests are added as well.

One minor logical change is dropping of tsmon.State after every
flush instead of resetting it. We were resetting all metrics
reported to the state, this is equivalent to dropping the state
and starting with a new fresh state.

R=chanli@chromium.org

Change-Id: I15380f4e102eaced28208f424e974e6f6feddd99
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5671720
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>