[Analysis] Track more confidence intervals for changepoints.
Rather than only track the 99% CI interval for changepoint position,
track a range of intervals. We identify a range of changepoint
position cumulative distribution function (CDF) points we want to
track the corresponding commit position for and store them in
Spanner in an efficient manner.

In a follow-up, we plan to expose the data to the UI via an
extension to the pRPC API.

This change also makes the following fixes/improvements to
confidence interval calculation:
- fixes a bug with the original 99% confidence intervals,
  where the lower bound was unnecessarily too conservative
  by one position.
- confidence interval is now over all source positions, not
  just those with test results.

BUG=None
TEST=Integration tests

Change-Id: I791aae2afafef14572d6de3b66d993d469fc5cf2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5671723
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>