[Analysis] Improve handling of unicode replacement character U+FFFD.
The unicode replacement character, often printed <?>, is printed
by some clients when given a character that is unknown or
not representable in unicode. It is useful for detecting unicode
errors.

It is also a valid unicode character in its own right, and is
uploaded by some clients as part of their own failure association
rules, possibly because the underlying error was not Unicode.

A recent occurrence saw bug filing attempt to file a bug for a
failure reason with such U+FFFD characters, which was blocked by
validation that was in place on the failure association rule syntax.

We wish to relax this syntax restriction, to allow encoding
U+FFFD characters in rules via escape sequences (\ufffd) but not
by including the characters directly.

It was validated that strconv.Unquote returns errors on Unicode
surrogate characters already so part of the existing validation
was redundant.

BUG=None
TEST=Integration tests

Change-Id: Id210e58b4d717395d7df4c058f22085ebd27a0b0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5680786
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>