[auth] Stop comparing to background.Context().
This doesn't work reliably when using e.g. timeouts (and it
was never promised to work at all).

Instead of trying to detect if a context is background
and switch to a different one, just "merge" two contexts.
This was already used in GetPerRPCCredentials, now use it
in GetRPCTransport as well.

This also somewhat relaxes paranoid checks for cross-request
transport sharing, since they are not very easy to do now.
In particular, a transport created from a user request can
now be used with a background context (it was forbidden
before). It still forbidden to create a transport within one
request handler, and then use it from another.

Finally, if a transport was created with a background
context, but used within a request context, it now uses
the request context auth data. This should be a noop, since
background context can only be used with credentials mode
that **do not** user auth data (attempting to create e.g.
GetRPCTransport(background, AsCredentialsForwarder) fails
right away, since it checks there's auth data in the
context).

R=iannucci@chromium.org

Change-Id: I51d0f551a15a332ab86e2e224186da50b4ce1d35
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5703721
Auto-Submit: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>