[swarming] Datastore entities for bot dimensions aggregation.
To serve GetBotDimensions(pool) API we need to store a map

  pool name => dimension key => [set of dimension values]

Python code stores this map in a bunch of entities: an entity
per pool, and one entity ("all") with union of all other entities.

When GetBotDimensions is called without a pool, it is supposed to
return dimensions of all bots visible to the user. Python code uses
"all" entry for this. And since "all" includes bots across ALL pools
(regardless of ACLs), Python implementation allows calling
GetBotDimensions without a pool only to users with server-global bot
listing permissions.

Go will do better. It will store the entire mapping in a single
entity (it should fit in compressed form, with lots of room to
spare). Then a set of "visible" dimensions can be relatively
efficiently calculated on the fly by first checking which pools
a user can see, and then returning a union of corresponding sets.
That way we don't need a weird ACL exception.

Loaded and parsed BotsDimensionsAggregation will be stored in memory.
To avoid reloading and reparsing it when nothing changes, there's
a tiny BotsDimensionsAggregationInfo entity with LastUpdate field. It
will be used to quickly check if there are changes to the stored
aggregation.

R=chanli@chromium.org

Change-Id: I64f24acd71bbae55221992dc69b3404247cf3349
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5709131
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>