resultdb: Add -log-level CLI flag for rdb tool
As-is, the rdb command hard-codes a logging level of "info". So local
invocations have no ability to quiet the amount of log lines.

This will surface the log-level via "-log-level" arg to let folks
change it, in case they only want to see warnings, errors, etc.

It will still default to "info", so this will be a no-op for anyone
that doesn't change it.

Bug: 41492688
Change-Id: I82ba7763230d96cec1d74f21521cb35bf8f74219
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5709412
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>