[ArtifactExporter] Add logging
I am a bit puzzled with the behavior of the artifact exporter.

When I rolled the checkpoint to canary, the task in canary got timed out. When I looked at the time out in the tasks, there are big time gaps between each write to checkpoint table. This suggested maybe something to do with downloading from CAS or exporting to BigQuery.

To my surprise, the Spanner insert doesn't take long, so perhaps it's not the problem with Spanner? But then why this only happens when we enabled checkpoint?

So add more logging to help us understand more.

Note: I already reverted the checkpoint CL, so adding the logging to the base first.

Bug:b/351046122
Change-Id: Ib5198b435b53a3d8a46d363fbd836fe0e19c1ec8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5719837
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>