[swarming] Recognize "swarming.terminate:1" tag.
It will be used soon to mark the termination tasks instead of
"terminate:1". The reason is that we'll reuse this tag for
deduplicating termination tasks created by TerminateBot, and it
would be nice if user-submitted tasks cannot set it (since this
can lead to weird interference with Swarming internals).

Forbidding setting "terminate" tag seems bad, since it is
a generic looking word that user task may actually want to use.
But "swarming.terminate" is pretty specific.

BUG=b/354247867
R=chanli@chromium.org

Change-Id: I53c12c47e3226518233bfb542ca72ea4159cb353
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5736469
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>