[testing/truth] Implement LineContext in a better way.
This removes the awkward method on `comparison.Func` and replaces it
with a more flexible system of Options which can be supplied to the
truth assertion/check functions.

The implementation allows users to implement their own Summary
modifier functions - the line context calculation is implemented in
terms of this generally available (and simple) affordance.

R=gregorynisbet@google.com

Bug: 319730573
Change-Id: I8f89b2d201d8ed0a997a7c935377b32abc7de517
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5740493
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>