[MILO] Separate default (frontend) and API services.
gae.py was not uploading API service because the service
was not following a standard directory structure anymore. (API was
in a subdirectory while the default module was occupying
the top-level folder).

https://cloud.google.com/appengine/docs/standard/configuration-files#examples

Move frontend and API into seperate folders to re-create structure
and avoid API service having to share the same entrypoint /
package as UI.

Delete pubsub service as nothing was using it.

BUG=b:349254870
TEST=make up-dev

Change-Id: If2c37c85131dfa7c33251f6986f8524ef7bd5704
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5740700
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>