[Common] Move bqutil to common
We have a few places that needs to write to BigQuery (LUCI Analysis, ResultDB, Bisection). In the near future, LUCI Tree Status also needs to export to bigquery.

Currently we are copying the code to different projects, which is not ideal. LUCI Common seems to be the best place for it.



Change-Id: Ib888e7019433fd4df000148b19b276f0ef4310b0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5762445
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>