[server] Initialize more of OTEL when running locally.
Before by default we were installing noop OTEL Provider
in dev mode (~= running locally) which effectively disabled
most of OTEL at an earliest entry point.

Now we install the real provider (together with the real "resource"
and the sampler and the batcher), but the noop exporter. That way
more of OTEL guts are exercised locally. This is important because
these guts to do not appear particularly trustworthy and without
running them locally we risk finding problems only in staging or
production (happened twice already).

One of such problems is essentially
https://github.com/open-telemetry/opentelemetry-go/issues/4476

Unfortunately, even with this CL we can't trigger this problem
locally, since the relevant OTEL code does nothing when running
not on GCP. There will be a separate CL that adds a test that
launches OTEL within an emulated GCP environment. Our existing
emulation (as done in `luci-auth context` seems to be enough
to trigger this particular bug).

One side effect of this change is that server.New(...) now
mutates global process state (since this is where OTEL keeps
its provider), which may be problematic for parallel tests.
There's only one test that calls server.New though, so we
can just isolate it.

R=iannucci@chromium.org
BUG=b/356517179

Change-Id: Icb92c1679bd8f3537990edd0d55bf83fbc1a715e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5774462
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>