[server] Add a test that launches empty server.
The primary purpose for now is to catch failures due to
bumps in OpenTelemetry dependencies. To hit all necessary code
paths we need to setup fake auth context (to trick OTEL into
thinking it runs on GCE).

This can in theory be extended somehow to run bigger integration
tests for the server, but likely a lot more design will be
necessary.

Extra care is taken to make sure `metadata.OnGCE` is called
for the first time when the GCE metadata server is already
faked. This first call happens from inside OpenTelemetry guts.
This is all super fragile, but there appears to be no other
way to trigger the OTEL error locally not on GCP.

R=iannucci@chromium.org
BUG=b/356517179

Change-Id: I060ed980e4a534405e96c19083dadd23091bbc63
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5775022
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>