[common] Enhance pubsub module to handle ignored messages.
For monitoring purposes it is useful to distinguish messages
on which we took some action versus those which were uninteresting
(e.g. replays of old messages, or messages we filtered out).

Define a pubsub.Ignore tag which allows a pub/sub handler to
indicate it is ignoring a message. When the message is ignored, the
handler returns status 204. This is based on a similar design to
the tq package.

BUG=b:349254870
TEST=Tests

Change-Id: I107af9fa2091d71319eb1d6451e7695844e6aeeb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5798891
Auto-Submit: Patrick Meiring <meiring@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>