[swarming] High-level logic of bot archive ingestion.
The overall intention is to eventually store built bot archives
(as references to blob entities) in the server config entity.
The config will have a mapping "canary => <blob>, stable => <blob>"
that RPC handlers would use to know when to instruct the bot to
update itself.

Right now the config entity is derived exclusively from configs
in LUCI Config. This CL changes it to be derived based on 3
inputs:
  * Configs in LUCI Config (including BotDeployment protos).
  * State in the CIPD: if e.g. "latest" version changes, we need
    to react to that.
  * EmbeddedBotSettings which eventually becomes config.json stored
    in the bot package. It is static now, but conceptually it is
    a separate input (i.e. if we even decide to add a new field to
    config.json, we'll need to rebuild bot archives).

Another consideration is that we want to know when the config
**actually** changes, because changes to the config git revision
or CIPD labels do not necessary mean the config is different. We
do not want to trash caches on unrelated infradata/config commits.

This leads to a scheme that uses deterministic build functions
and digests to update parts that change when they actually change.

The final config digest is now essentially:
  hash(
    hash(LUCI Config files),
    hash(canary CIPD package data, bot_config.py, settings.json),
    hash(stable CIPD package data, bot_config.py, settings.json),
  )

Finally, make sure Fetched timestamp changes only monotonically
and only when configs actually change. This would allow to do
some optimizations in the config caching code (we'd know if
the currently cached config is stale or not by comparing
timestamps).

R=chanli@chromium.org
BUG=b/355013257

Change-Id: Ibb9fbd13f17777331f1310ccff4bb51c7e0c607f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5805255
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>